Add hook to create account user in local storage
