+++ src/components/UI/Account/features/AccountDialog/AccountDialog.tsx
@@ -8,7 +8,7 @@ import { Field, Formik } from 'formik';
 import { ERROR_MESSAGE_GENERAL } from '../../../../../constants';
 import { TYPE_OF_ACCOUNTS } from '../../../../../globalInterface';
 import { SystemStateEnum } from '../../../../../enums';
-import { useAppDispatch, useAppSelector } from '../../../../../redux/hooks';
+import { useAppSelector } from '../../../../../redux/hooks';
 import { CreateAccountMutationProps, ModifyAccountMutationProps } from '../../../../../redux/slices/Accounts/interface';
 import { useNotification } from '../../../../../hooks/useNotification';
 import { useCurrencyField } from '../../../../Other/CurrencyField/useCurrencyField';
@@ -26,10 +26,7 @@ import { LoadingSpinner } from '../../../LoadingSpinner';
 import { CurrencyField } from '../../../../Other';
 import { useCreateAccountMutation, useModifyAccountMutation } from '../../../../../redux/slices/Accounts/actions';
 import { useGuestUser } from '../../../../../hooks/useGuestUser/useGuestUser';
-import {
-  addToLocalStorage, formatAccountsForLocalStorage, formatSingleAccount,
-} from '../../../../../utils';
-import { updateAccounts, updateAccountsLocalStorage } from '../../../../../redux/slices/Accounts/accounts.slice';
+import { useAccount } from '../../../../../hooks/useAccount';
 
 const initialValuesCreateAccount: CreateAccountInitialValues = {
   title: '',
@@ -44,8 +41,8 @@ const AccountDialog = ({
   accountAction,
   account,
 }: AccountDialogProps) => {
-  const dispatch = useAppDispatch();
   const { isGuestUser } = useGuestUser();
+  const { createAccountGuestUser } = useAccount();
   const { updateAmount, initialAmount } = useCurrencyField();
   const [createAccountMutation, { isLoading: isLoadingCreateAccount }] = useCreateAccountMutation();
   const [modifyAccountMutation, { isLoading: isLoadingModifyAccount }] = useModifyAccountMutation();
@@ -53,8 +50,6 @@ const AccountDialog = ({
   const { updateGlobalNotification } = useNotification();
   const userReduxState = useAppSelector((state) => state.user);
   const bearerToken = userReduxState.userInfo?.bearerToken as string;
-  const accountsLocalStorage = useAppSelector((state) => state.accounts.accountsLocalStorage);
-  const accountsUI = useAppSelector((state) => state.accounts.accounts);
 
   // Copying constant because it is readyonly
   const typeAccounts = [...TYPE_OF_ACCOUNTS];
@@ -79,15 +74,8 @@ const AccountDialog = ({
       const createAccountValues: CreateAccount = { ...values, amount: amountNumber, color: 'black' };
       const createAccountMutationProps: CreateAccountMutationProps = { values: createAccountValues, bearerToken };
 
-      if (isGuestUser && accountsLocalStorage && accountsUI) {
-        const formattedAccount = formatAccountsForLocalStorage(createAccountValues);
-        const newAccountUI = formatSingleAccount(formattedAccount);
-        const newAccountsUI = [...accountsUI, newAccountUI];
-        const newAccounts = [...accountsLocalStorage, formattedAccount];
-
-        addToLocalStorage({ newInfo: newAccounts, prop: 'accounts' });
-        dispatch(updateAccounts(newAccountsUI));
-        dispatch(updateAccountsLocalStorage(newAccounts));
+      if (isGuestUser) {
+        createAccountGuestUser(createAccountValues);
 
         // Show success notification
         updateGlobalNotification({
